Crate rusoto_elasticache

Amazon ElastiCache
Amazon ElastiCache is a web service that makes it easier to set up, operate, and scale a distributed cache in the cloud.

With ElastiCache, customers get all of the benefits of a high-performance, in-memory cache with less of the administrative burden involved in launching and managing a distributed cache. The service makes setup, scaling, and cluster failure handling much simpler than in a self-managed cache deployment.

In addition, through integration with Amazon CloudWatch, customers get enhanced visibility into the key performance statistics associated with their cache and can receive alarms if a part of their cache runs hot.

If you're using the service, you're probably looking for ElastiCacheClient and ElastiCache.

Structs

AddTagsToResourceMessage
Represents the input of an AddTagsToResource operation.

AllowedNodeTypeModificationsMessage
Represents the allowed node types you can use to modify your cluster or replication group.

AuthorizeCacheSecurityGroupIngressMessage
Represents the input of an AuthorizeCacheSecurityGroupIngress operation.

AuthorizeCacheSecurityGroupIngressResult

AvailabilityZone
Describes an Availability Zone in which the cluster is launched.

BatchApplyUpdateActionMessage

BatchStopUpdateActionMessage

CacheCluster
Contains all of the attributes of a specific cluster.

CacheClusterMessage
Represents the output of a DescribeCacheClusters operation.

CacheEngineVersion
Provides all of the details about a particular cache engine version.

CacheEngineVersionMessage
Represents the output of a DescribeCacheEngineVersions operation.

CacheNode

Represents an individual cache node within a cluster. Each cache node runs its own instance of the cluster's protocol-compliant caching software - either Memcached or Redis.

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

CacheNodeTypeSpecificParameter
A parameter that has a different value for each cache node type it is applied to. For example, in a Redis cluster, a cache.m1.large cache node type would have a larger maxmemory value than a cache.m1.small type.

CacheNodeTypeSpecificValue
A value that applies only to a certain cache node type.

CacheParameterGroup
Represents the output of a CreateCacheParameterGroup operation.

CacheParameterGroupDetails
Represents the output of a DescribeCacheParameters operation.

CacheParameterGroupNameMessage

Represents the output of one of the following operations:

ModifyCacheParameterGroup

ResetCacheParameterGroup

CacheParameterGroupStatus
Status of the cache parameter group.

CacheParameterGroupsMessage
Represents the output of a DescribeCacheParameterGroups operation.

CacheSecurityGroup

Represents the output of one of the following operations:

AuthorizeCacheSecurityGroupIngress

CreateCacheSecurityGroup

RevokeCacheSecurityGroupIngress

CacheSecurityGroupMembership
Represents a cluster's status within a particular cache security group.

CacheSecurityGroupMessage
Represents the output of a DescribeCacheSecurityGroups operation.

CacheSubnetGroup

Represents the output of one of the following operations:

CreateCacheSubnetGroup

ModifyCacheSubnetGroup

CacheSubnetGroupMessage
Represents the output of a DescribeCacheSubnetGroups operation.

ConfigureShard
Node group (shard) configuration options when adding or removing replicas. Each node group (shard) configuration has the following members: NodeGroupId, NewReplicaCount, and PreferredAvailabilityZones.

CopySnapshotMessage
Represents the input of a CopySnapshotMessage operation.

CopySnapshotResult

CreateCacheClusterMessage
Represents the input of a CreateCacheCluster operation.

CreateCacheClusterResult

CreateCacheParameterGroupMessage
Represents the input of a CreateCacheParameterGroup operation.

CreateCacheParameterGroupResult

CreateCacheSecurityGroupMessage
Represents the input of a CreateCacheSecurityGroup operation.

CreateCacheSecurityGroupResult

CreateCacheSubnetGroupMessage
Represents the input of a CreateCacheSubnetGroup operation.

CreateCacheSubnetGroupResult

CreateReplicationGroupMessage
Represents the input of a CreateReplicationGroup operation.

CreateReplicationGroupResult

CreateSnapshotMessage
Represents the input of a CreateSnapshot operation.

CreateSnapshotResult

DecreaseReplicaCountMessage

DecreaseReplicaCountResult

DeleteCacheClusterMessage
Represents the input of a DeleteCacheCluster operation.

DeleteCacheClusterResult

DeleteCacheParameterGroupMessage
Represents the input of a DeleteCacheParameterGroup operation.

DeleteCacheSecurityGroupMessage
Represents the input of a DeleteCacheSecurityGroup operation.

DeleteCacheSubnetGroupMessage
Represents the input of a DeleteCacheSubnetGroup operation.

DeleteReplicationGroupMessage
Represents the input of a DeleteReplicationGroup operation.

DeleteReplicationGroupResult

DeleteSnapshotMessage
Represents the input of a DeleteSnapshot operation.

DeleteSnapshotResult

DescribeCacheClustersMessage
Represents the input of a DescribeCacheClusters operation.

DescribeCacheEngineVersionsMessage
Represents the input of a DescribeCacheEngineVersions operation.

DescribeCacheParameterGroupsMessage
Represents the input of a DescribeCacheParameterGroups operation.

DescribeCacheParametersMessage
Represents the input of a DescribeCacheParameters operation.

DescribeCacheSecurityGroupsMessage
Represents the input of a DescribeCacheSecurityGroups operation.

DescribeCacheSubnetGroupsMessage
Represents the input of a DescribeCacheSubnetGroups operation.

DescribeEngineDefaultParametersMessage
Represents the input of a DescribeEngineDefaultParameters operation.

DescribeEngineDefaultParametersResult

DescribeEventsMessage
Represents the input of a DescribeEvents operation.

DescribeReplicationGroupsMessage
Represents the input of a DescribeReplicationGroups operation.

DescribeReservedCacheNodesMessage
Represents the input of a DescribeReservedCacheNodes operation.

DescribeReservedCacheNodesOfferingsMessage
Represents the input of a DescribeReservedCacheNodesOfferings operation.

DescribeServiceUpdatesMessage

DescribeSnapshotsListMessage
Represents the output of a DescribeSnapshots operation.

DescribeSnapshotsMessage
Represents the input of a DescribeSnapshotsMessage operation.

DescribeUpdateActionsMessage

EC2SecurityGroup
Provides ownership and status information for an Amazon EC2 security group.

ElastiCacheClient
A client for the Amazon ElastiCache API.

Endpoint
Represents the information required for client programs to connect to a cache node.

EngineDefaults
Represents the output of a DescribeEngineDefaultParameters operation.

Event
Represents a single occurrence of something interesting within the system. Some examples of events are creating a cluster, adding or removing a cache node, or rebooting a node.

EventsMessage
Represents the output of a DescribeEvents operation.

IncreaseReplicaCountMessage

IncreaseReplicaCountResult

ListAllowedNodeTypeModificationsMessage
The input parameters for the ListAllowedNodeTypeModifications operation.

ListTagsForResourceMessage
The input parameters for the ListTagsForResource operation.

ModifyCacheClusterMessage
Represents the input of a ModifyCacheCluster operation.

ModifyCacheClusterResult

ModifyCacheParameterGroupMessage
Represents the input of a ModifyCacheParameterGroup operation.

ModifyCacheSubnetGroupMessage
Represents the input of a ModifyCacheSubnetGroup operation.

ModifyCacheSubnetGroupResult

ModifyReplicationGroupMessage
Represents the input of a ModifyReplicationGroups operation.

ModifyReplicationGroupResult

ModifyReplicationGroupShardConfigurationMessage
Represents the input for a ModifyReplicationGroupShardConfiguration operation.

ModifyReplicationGroupShardConfigurationResult

NodeGroup
Represents a collection of cache nodes in a replication group. One node in the node group is the read/write primary node. All the other nodes are read-only Replica nodes.

NodeGroupConfiguration
Node group (shard) configuration options. Each node group (shard) configuration has the following: Slots, PrimaryAvailabilityZone, ReplicaAvailabilityZones, ReplicaCount.

NodeGroupMember
Represents a single node within a node group (shard).

NodeGroupMemberUpdateStatus
The status of the service update on the node group member

NodeGroupUpdateStatus
The status of the service update on the node group

NodeSnapshot
Represents an individual cache node in a snapshot of a cluster.

NotificationConfiguration
Describes a notification topic and its status. Notification topics are used for publishing ElastiCache events to subscribers using Amazon Simple Notification Service (SNS).

Parameter
Describes an individual setting that controls some aspect of ElastiCache behavior.

ParameterNameValue
Describes a name-value pair that is used to update the value of a parameter.

PendingModifiedValues
A group of settings that are applied to the cluster in the future, or that are currently being applied.

ProcessedUpdateAction
Update action that has been processed for the corresponding apply/stop request

PurchaseReservedCacheNodesOfferingMessage
Represents the input of a PurchaseReservedCacheNodesOffering operation.

PurchaseReservedCacheNodesOfferingResult

RebootCacheClusterMessage
Represents the input of a RebootCacheCluster operation.

RebootCacheClusterResult

RecurringCharge
Contains the specific price and frequency of a recurring charges for a reserved cache node, or for a reserved cache node offering.

RemoveTagsFromResourceMessage
Represents the input of a RemoveTagsFromResource operation.

ReplicationGroup
Contains all of the attributes of a specific Redis replication group.

ReplicationGroupMessage
Represents the output of a DescribeReplicationGroups operation.

ReplicationGroupPendingModifiedValues
The settings to be applied to the Redis replication group, either immediately or during the next maintenance window.

ReservedCacheNode
Represents the output of a PurchaseReservedCacheNodesOffering operation.

ReservedCacheNodeMessage
Represents the output of a DescribeReservedCacheNodes operation.

ReservedCacheNodesOffering
Describes all of the attributes of a reserved cache node offering.

ReservedCacheNodesOfferingMessage
Represents the output of a DescribeReservedCacheNodesOfferings operation.

ResetCacheParameterGroupMessage
Represents the input of a ResetCacheParameterGroup operation.

ReshardingConfiguration
A list of PreferredAvailabilityZones objects that specifies the configuration of a node group in the resharded cluster.

ReshardingStatus
The status of an online resharding operation.

RevokeCacheSecurityGroupIngressMessage
Represents the input of a RevokeCacheSecurityGroupIngress operation.

RevokeCacheSecurityGroupIngressResult

SecurityGroupMembership
Represents a single cache security group and its status.

ServiceUpdate
An update that you can apply to your Redis clusters.

ServiceUpdatesMessage

SlotMigration
Represents the progress of an online resharding operation.

Snapshot
Represents a copy of an entire Redis cluster as of the time when the snapshot was taken.

Subnet
Represents the subnet associated with a cluster. This parameter refers to subnets defined in Amazon Virtual Private Cloud (Amazon VPC) and used with ElastiCache.

Tag
A cost allocation Tag that can be added to an ElastiCache cluster or replication group. Tags are composed of a Key/Value pair. A tag with a null Value is permitted.

TagListMessage
Represents the output from the AddTagsToResource, ListTagsForResource, and RemoveTagsFromResource operations.

TestFailoverMessage

TestFailoverResult

TimeRangeFilter
Filters update actions from the service updates that are in available status during the time range.

UnprocessedUpdateAction
Update action that has failed to be processed for the corresponding apply/stop request

UpdateAction
The status of the service update for a specific replication group

UpdateActionResultsMessage

UpdateActionsMessage

Enums

AddTagsToResourceError
Errors returned by AddTagsToResource

AuthorizeCacheSecurityGroupIngressError
Errors returned by AuthorizeCacheSecurityGroupIngress

BatchApplyUpdateActionError
Errors returned by BatchApplyUpdateAction

BatchStopUpdateActionError
Errors returned by BatchStopUpdateAction

CopySnapshotError
Errors returned by CopySnapshot

CreateCacheClusterError
Errors returned by CreateCacheCluster

CreateCacheParameterGroupError
Errors returned by CreateCacheParameterGroup

CreateCacheSecurityGroupError
Errors returned by CreateCacheSecurityGroup

CreateCacheSubnetGroupError
Errors returned by CreateCacheSubnetGroup

CreateReplicationGroupError
Errors returned by CreateReplicationGroup

CreateSnapshotError
Errors returned by CreateSnapshot

DecreaseReplicaCountError
Errors returned by DecreaseReplicaCount

DeleteCacheClusterError
Errors returned by DeleteCacheCluster

DeleteCacheParameterGroupError
Errors returned by DeleteCacheParameterGroup

DeleteCacheSecurityGroupError
Errors returned by DeleteCacheSecurityGroup

DeleteCacheSubnetGroupError
Errors returned by DeleteCacheSubnetGroup

DeleteReplicationGroupError
Errors returned by DeleteReplicationGroup

DeleteSnapshotError
Errors returned by DeleteSnapshot

DescribeCacheClustersError
Errors returned by DescribeCacheClusters

DescribeCacheEngineVersionsError
Errors returned by DescribeCacheEngineVersions

DescribeCacheParameterGroupsError
Errors returned by DescribeCacheParameterGroups

DescribeCacheParametersError
Errors returned by DescribeCacheParameters

DescribeCacheSecurityGroupsError
Errors returned by DescribeCacheSecurityGroups

DescribeCacheSubnetGroupsError
Errors returned by DescribeCacheSubnetGroups

DescribeEngineDefaultParametersError
Errors returned by DescribeEngineDefaultParameters

DescribeEventsError
Errors returned by DescribeEvents

DescribeReplicationGroupsError
Errors returned by DescribeReplicationGroups

DescribeReservedCacheNodesError
Errors returned by DescribeReservedCacheNodes

DescribeReservedCacheNodesOfferingsError
Errors returned by DescribeReservedCacheNodesOfferings

DescribeServiceUpdatesError
Errors returned by DescribeServiceUpdates

DescribeSnapshotsError
Errors returned by DescribeSnapshots

DescribeUpdateActionsError
Errors returned by DescribeUpdateActions

IncreaseReplicaCountError
Errors returned by IncreaseReplicaCount

ListAllowedNodeTypeModificationsError
Errors returned by ListAllowedNodeTypeModifications

ListTagsForResourceError
Errors returned by ListTagsForResource

ModifyCacheClusterError
Errors returned by ModifyCacheCluster

ModifyCacheParameterGroupError
Errors returned by ModifyCacheParameterGroup

ModifyCacheSubnetGroupError
Errors returned by ModifyCacheSubnetGroup

ModifyReplicationGroupError
Errors returned by ModifyReplicationGroup

ModifyReplicationGroupShardConfigurationError
Errors returned by ModifyReplicationGroupShardConfiguration

PurchaseReservedCacheNodesOfferingError
Errors returned by PurchaseReservedCacheNodesOffering

RebootCacheClusterError
Errors returned by RebootCacheCluster

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

ResetCacheParameterGroupError
Errors returned by ResetCacheParameterGroup

RevokeCacheSecurityGroupIngressError
Errors returned by RevokeCacheSecurityGroupIngress

TestFailoverError
Errors returned by TestFailover

Traits

ElastiCache
Trait representing the capabilities of the Amazon ElastiCache API. Amazon ElastiCache clients implement this trait.