Enum rusoto_elasticache::CreateReplicationGroupError

pub enum CreateReplicationGroupError {
    CacheClusterNotFoundFault(String),
    CacheParameterGroupNotFoundFault(String),
    CacheSecurityGroupNotFoundFault(String),
    CacheSubnetGroupNotFoundFault(String),
    ClusterQuotaForCustomerExceededFault(String),
    InsufficientCacheClusterCapacityFault(String),
    InvalidCacheClusterStateFault(String),
    InvalidParameterCombination(String),
    InvalidParameterValue(String),
    InvalidVPCNetworkStateFault(String),
    NodeGroupsPerReplicationGroupQuotaExceededFault(String),
    NodeQuotaForClusterExceededFault(String),
    NodeQuotaForCustomerExceededFault(String),
    ReplicationGroupAlreadyExistsFault(String),
    TagQuotaPerResourceExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateReplicationGroup

Variants
CacheClusterNotFoundFault(String)

The requested cluster ID does not refer to an existing cluster.

CacheParameterGroupNotFoundFault(String)

The requested cache parameter group name does not refer to an existing cache parameter group.

CacheSecurityGroupNotFoundFault(String)

The requested cache security group name does not refer to an existing cache security group.

CacheSubnetGroupNotFoundFault(String)

The requested cache subnet group name does not refer to an existing cache subnet group.

ClusterQuotaForCustomerExceededFault(String)

The request cannot be processed because it would exceed the allowed number of clusters per customer.

InsufficientCacheClusterCapacityFault(String)

The requested cache node type is not available in the specified Availability Zone.

InvalidCacheClusterStateFault(String)

The requested cluster is not in the available state.

InvalidParameterCombination(String)

Two or more incompatible parameters were specified.

InvalidParameterValue(String)

The value for a parameter is invalid.

InvalidVPCNetworkStateFault(String)

The VPC network is in an invalid state.

NodeGroupsPerReplicationGroupQuotaExceededFault(String)

The request cannot be processed because it would exceed the maximum allowed number of node groups (shards) in a single replication group. The default maximum is 15

NodeQuotaForClusterExceededFault(String)

The request cannot be processed because it would exceed the allowed number of cache nodes in a single cluster.

NodeQuotaForCustomerExceededFault(String)

The request cannot be processed because it would exceed the allowed number of cache nodes per customer.

ReplicationGroupAlreadyExistsFault(String)

The specified replication group already exists.

TagQuotaPerResourceExceeded(String)

The request cannot be processed because it would cause the resource to have more than the allowed number of tags. The maximum number of tags permitted on a resource is 50.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateReplicationGroupError

pub fn from_body(body: &str) -> CreateReplicationGroupError