Struct rusoto_elasticache::CacheNode

pub struct CacheNode {
    pub cache_node_create_time: Option<String>,
    pub cache_node_id: Option<String>,
    pub cache_node_status: Option<String>,
    pub customer_availability_zone: Option<String>,
    pub endpoint: Option<Endpoint>,
    pub parameter_group_status: Option<String>,
    pub source_cache_node_id: Option<String>,
}

Represents an individual cache node within a cache cluster. Each cache node runs its own instance of the cluster's protocol-compliant caching software - either Memcached or Redis.

Valid node types are as follows:

General purpose:

Current generation: cache.t2.micro, cache.t2.small, cache.t2.medium, cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge, cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

Previous generation: cache.t1.micro, cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

Compute optimized: cache.c1.xlarge

Memory optimized:

Current generation: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Previous generation: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

Notes:

All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC).

Redis backup/restore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backup/restore is supported on Redis (cluster mode enabled) T2 instances.

Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances.

For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis.

Fields
cache_node_create_time: Option<String>

The date and time when the cache node was created.

cache_node_id: Option<String>

The cache node identifier. A node ID is a numeric identifier (0001, 0002, etc.). The combination of cluster ID and node ID uniquely identifies every cache node used in a customer's AWS account.

cache_node_status: Option<String>

The current state of this cache node.

customer_availability_zone: Option<String>

The Availability Zone where this node was created and now resides.

endpoint: Option<Endpoint>

The hostname for connecting to this cache node.

parameter_group_status: Option<String>

The status of the parameter group applied to this cache node.

source_cache_node_id: Option<String>

The ID of the primary node to which this read replica node is synchronized. If this field is empty, this node is not associated with a primary cache cluster.