Struct rusoto_elasticache::CreateReplicationGroupMessage

pub struct CreateReplicationGroupMessage {
    pub auth_token: Option<String>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub automatic_failover_enabled: Option<bool>,
    pub cache_node_type: Option<String>,
    pub cache_parameter_group_name: Option<String>,
    pub cache_security_group_names: Option<Vec<String>>,
    pub cache_subnet_group_name: Option<String>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub node_group_configuration: Option<Vec<NodeGroupConfiguration>>,
    pub notification_topic_arn: Option<String>,
    pub num_cache_clusters: Option<i64>,
    pub num_node_groups: Option<i64>,
    pub port: Option<i64>,
    pub preferred_cache_cluster_a_zs: Option<Vec<String>>,
    pub preferred_maintenance_window: Option<String>,
    pub primary_cluster_id: Option<String>,
    pub replicas_per_node_group: Option<i64>,
    pub replication_group_description: String,
    pub replication_group_id: String,
    pub security_group_ids: Option<Vec<String>>,
    pub snapshot_arns: Option<Vec<String>>,
    pub snapshot_name: Option<String>,
    pub snapshot_retention_limit: Option<i64>,
    pub snapshot_window: Option<String>,
    pub tags: Option<Vec<Tag>>,
}

Represents the input of a CreateReplicationGroup operation.

Fields
auth_token: Option<String>

Reserved parameter. The password used to access a password protected server.

Password constraints:

Must be only printable ASCII characters.

Must be at least 16 characters and no more than 128 characters in length.

Cannot contain any of the following characters: '/', '"', or "@".

For more information, see AUTH password at Redis.

auto_minor_version_upgrade: Option<bool>

This parameter is currently disabled.

automatic_failover_enabled: Option<bool>

Specifies whether a read-only replica is automatically promoted to read/write primary if the existing primary fails.

If true, Multi-AZ is enabled for this replication group. If false, Multi-AZ is disabled for this replication group.

AutomaticFailoverEnabled must be enabled for Redis (cluster mode enabled) replication groups.

Default: false

ElastiCache Multi-AZ replication groups is not supported on:

Redis versions earlier than 2.8.6.

Redis (cluster mode disabled): T1 and T2 node types.

Redis (cluster mode enabled): T2 node types.

cache_node_type: Option<String>

The compute and memory capacity of the nodes in the node group (shard).

Valid node types are as follows:

General purpose:

Current generation: cache.t2.micro, cache.t2.small, cache.t2.medium, cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge, cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

Previous generation: cache.t1.micro, cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

Compute optimized: cache.c1.xlarge

Memory optimized:

Current generation: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Previous generation: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

Notes:

All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC).

Redis backup/restore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backup/restore is supported on Redis (cluster mode enabled) T2 instances.

Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances.

For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis.

cache_parameter_group_name: Option<String>

The name of the parameter group to associate with this replication group. If this argument is omitted, the default cache parameter group for the specified engine is used.

If you are running Redis version 3.2.4 or later, only one node group (shard), and want to use a default parameter group, we recommend that you specify the parameter group by name.

To create a Redis (cluster mode disabled) replication group, use CacheParameterGroupName=default.redis3.2.

To create a Redis (cluster mode enabled) replication group, use CacheParameterGroupName=default.redis3.2.cluster.on.

cache_security_group_names: Option<Vec<String>>

A list of cache security group names to associate with this replication group.

cache_subnet_group_name: Option<String>

The name of the cache subnet group to be used for the replication group.

If you're going to launch your cluster in an Amazon VPC, you need to create a subnet group before you start creating a cluster. For more information, see Subnets and Subnet Groups.

engine: Option<String>

The name of the cache engine to be used for the cache clusters in this replication group.

engine_version: Option<String>

The version number of the cache engine to be used for the cache clusters in this replication group. To view the supported cache engine versions, use the DescribeCacheEngineVersions operation.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version) in the ElastiCache User Guide, but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cache cluster or replication group and create it anew with the earlier engine version.

node_group_configuration: Option<Vec<NodeGroupConfiguration>>

A list of node group (shard) configuration options. Each node group (shard) configuration has the following: Slots, PrimaryAvailabilityZone, ReplicaAvailabilityZones, ReplicaCount.

If you're creating a Redis (cluster mode disabled) or a Redis (cluster mode enabled) replication group, you can use this parameter to individually configure each node group (shard), or you can omit this parameter.

notification_topic_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

The Amazon SNS topic owner must be the same as the cache cluster owner.

num_cache_clusters: Option<i64>

The number of clusters this replication group initially has.

This parameter is not used if there is more than one node group (shard). You should use ReplicasPerNodeGroup instead.

If AutomaticFailoverEnabled is true, the value of this parameter must be at least 2. If AutomaticFailoverEnabled is false you can omit this parameter (it will default to 1), or you can explicitly set it to a value between 2 and 6.

The maximum permitted value for NumCacheClusters is 6 (primary plus 5 replicas).

num_node_groups: Option<i64>

An optional parameter that specifies the number of node groups (shards) for this Redis (cluster mode enabled) replication group. For Redis (cluster mode disabled) either omit this parameter or set it to 1.

Default: 1

port: Option<i64>

The port number on which each member of the replication group accepts connections.

preferred_cache_cluster_a_zs: Option<Vec<String>>

A list of EC2 Availability Zones in which the replication group's cache clusters are created. The order of the Availability Zones in the list is the order in which clusters are allocated. The primary cluster is created in the first AZ in the list.

This parameter is not used if there is more than one node group (shard). You should use NodeGroupConfiguration instead.

If you are creating your replication group in an Amazon VPC (recommended), you can only locate cache clusters in Availability Zones associated with the subnets in the selected subnet group.

The number of Availability Zones listed must equal the value of NumCacheClusters.

Default: system chosen Availability Zones.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the cache cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

primary_cluster_id: Option<String>

The identifier of the cache cluster that serves as the primary for this replication group. This cache cluster must already exist and have a status of available.

This parameter is not required if NumCacheClusters, NumNodeGroups, or ReplicasPerNodeGroup is specified.

replicas_per_node_group: Option<i64>

An optional parameter that specifies the number of replica nodes in each node group (shard). Valid values are 0 to 5.

replication_group_description: String

A user-created description for the replication group.

replication_group_id: String

The replication group identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

security_group_ids: Option<Vec<String>>

One or more Amazon VPC security groups associated with this replication group.

Use this parameter only when you are creating a replication group in an Amazon Virtual Private Cloud (Amazon VPC).

snapshot_arns: Option<Vec<String>>

A list of Amazon Resource Names (ARN) that uniquely identify the Redis RDB snapshot files stored in Amazon S3. The snapshot files are used to populate the new replication group. The Amazon S3 object name in the ARN cannot contain any commas. The new replication group will have the number of node groups (console: shards) specified by the parameter NumNodeGroups or the number of node groups configured by NodeGroupConfiguration regardless of the number of ARNs specified here.

This parameter is only valid if the Engine parameter is redis.

Example of an Amazon S3 ARN: arn:aws:s3:::my_bucket/snapshot1.rdb

snapshot_name: Option<String>

The name of a snapshot from which to restore data into the new replication group. The snapshot status changes to restoring while the new replication group is being created.

This parameter is only valid if the Engine parameter is redis.

snapshot_retention_limit: Option<i64>

The number of days for which ElastiCache retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

This parameter is only valid if the Engine parameter is redis.

Default: 0 (i.e., automatic backups are disabled for this cache cluster).

snapshot_window: Option<String>

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

tags: Option<Vec<Tag>>

A list of cost allocation tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value.