Struct rusoto_elasticache::DescribeReservedCacheNodesOfferingsMessage

pub struct DescribeReservedCacheNodesOfferingsMessage {
    pub cache_node_type: Option<String>,
    pub duration: Option<String>,
    pub marker: Option<String>,
    pub max_records: Option<i64>,
    pub offering_type: Option<String>,
    pub product_description: Option<String>,
    pub reserved_cache_nodes_offering_id: Option<String>,
}

Represents the input of a DescribeReservedCacheNodesOfferings operation.

Fields
cache_node_type: Option<String>

The cache node type filter value. Use this parameter to show only the available offerings matching the specified cache node type.

Valid node types are as follows:

General purpose:

Current generation: cache.t2.micro, cache.t2.small, cache.t2.medium, cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge, cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

Previous generation: cache.t1.micro, cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

Compute optimized: cache.c1.xlarge

Memory optimized:

Current generation: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Previous generation: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

Notes:

All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC).

Redis backup/restore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backup/restore is supported on Redis (cluster mode enabled) T2 instances.

Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances.

For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis.

duration: Option<String>

Duration filter value, specified in years or seconds. Use this parameter to show only reservations for a given duration.

Valid Values: 1 | 3 | 31536000 | 94608000

marker: Option<String>

An optional marker returned from a prior request. Use this marker for pagination of results from this operation. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

max_records: Option<i64>

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: minimum 20; maximum 100.

offering_type: Option<String>

The offering type filter value. Use this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Light Utilization"|"Medium Utilization"|"Heavy Utilization"

product_description: Option<String>

The product description filter value. Use this parameter to show only the available offerings matching the specified product description.

reserved_cache_nodes_offering_id: Option<String>

The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706