Struct rusoto_elasticache::ElastiCacheClient

pub struct ElastiCacheClient { /* fields omitted */ }

A client for the Amazon ElastiCache API.

Methods

impl ElastiCacheClient

pub fn new(region: Region) -> ElastiCacheClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ElastiCacheClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,