Struct rusoto_elasticache::ModifyCacheClusterMessage

pub struct ModifyCacheClusterMessage {
    pub az_mode: Option<String>,
    pub apply_immediately: Option<bool>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub cache_cluster_id: String,
    pub cache_node_ids_to_remove: Option<Vec<String>>,
    pub cache_node_type: Option<String>,
    pub cache_parameter_group_name: Option<String>,
    pub cache_security_group_names: Option<Vec<String>>,
    pub engine_version: Option<String>,
    pub new_availability_zones: Option<Vec<String>>,
    pub notification_topic_arn: Option<String>,
    pub notification_topic_status: Option<String>,
    pub num_cache_nodes: Option<i64>,
    pub preferred_maintenance_window: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub snapshot_retention_limit: Option<i64>,
    pub snapshot_window: Option<String>,
}

Represents the input of a ModifyCacheCluster operation.

Fields
az_mode: Option<String>

Specifies whether the new nodes in this Memcached cache cluster are all created in a single Availability Zone or created across multiple Availability Zones.

Valid values: single-az | cross-az.

This option is only supported for Memcached cache clusters.

You cannot specify single-az if the Memcached cache cluster already has cache nodes in different Availability Zones. If cross-az is specified, existing Memcached nodes remain in their current Availability Zone.

Only newly created nodes are located in different Availability Zones. For instructions on how to move existing Memcached nodes to different Availability Zones, see the Availability Zone Considerations section of Cache Node Considerations for Memcached.

apply_immediately: Option<bool>

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the cache cluster.

If false, changes to the cache cluster are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

If you perform a ModifyCacheCluster before a pending modification is applied, the pending modification is replaced by the newer modification.

Valid values: true | false

Default: false

auto_minor_version_upgrade: Option<bool>

This parameter is currently disabled.

cache_cluster_id: String

The cache cluster identifier. This value is stored as a lowercase string.

cache_node_ids_to_remove: Option<Vec<String>>

A list of cache node IDs to be removed. A node ID is a numeric identifier (0001, 0002, etc.). This parameter is only valid when NumCacheNodes is less than the existing number of cache nodes. The number of cache node IDs supplied in this parameter must match the difference between the existing number of cache nodes in the cluster or pending cache nodes, whichever is greater, and the value of NumCacheNodes in the request.

For example: If you have 3 active cache nodes, 7 pending cache nodes, and the number of cache nodes in this ModifyCacheCluser call is 5, you must list 2 (7 - 5) cache node IDs to remove.

cache_node_type: Option<String>

A valid cache node type that you want to scale this cache cluster up to.

cache_parameter_group_name: Option<String>

The name of the cache parameter group to apply to this cache cluster. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

cache_security_group_names: Option<Vec<String>>

A list of cache security group names to authorize on this cache cluster. This change is asynchronously applied as soon as possible.

You can use this parameter only with clusters that are created outside of an Amazon Virtual Private Cloud (Amazon VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must not be "Default".

engine_version: Option<String>

The upgraded version of the cache engine to be run on the cache nodes.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cache cluster and create it anew with the earlier engine version.

new_availability_zones: Option<Vec<String>>

The list of Availability Zones where the new Memcached cache nodes are created.

This parameter is only valid when NumCacheNodes in the request is greater than the sum of the number of active cache nodes and the number of cache nodes pending creation (which may be zero). The number of Availability Zones supplied in this list must match the cache nodes being added in this request.

This option is only supported on Memcached clusters.

Scenarios:

Scenario 1: You have 3 active nodes and wish to add 2 nodes. Specify NumCacheNodes=5 (3 + 2) and optionally specify two Availability Zones for the two new nodes.

Scenario 2: You have 3 active nodes and 2 nodes pending creation (from the scenario 1 call) and want to add 1 more node. Specify NumCacheNodes=6 ((3 + 2) + 1) and optionally specify an Availability Zone for the new node.

Scenario 3: You want to cancel all pending operations. Specify NumCacheNodes=3 to cancel all pending operations.

The Availability Zone placement of nodes pending creation cannot be modified. If you wish to cancel any nodes pending creation, add 0 nodes by setting NumCacheNodes to the number of current nodes.

If cross-az is specified, existing Memcached nodes remain in their current Availability Zone. Only newly created nodes can be located in different Availability Zones. For guidance on how to move existing Memcached nodes to different Availability Zones, see the Availability Zone Considerations section of Cache Node Considerations for Memcached.

Impact of new add/remove requests upon pending requests

Scenario-1

Pending Action: Delete

New Request: Delete

Result: The new delete, pending or immediate, replaces the pending delete.

Scenario-2

Pending Action: Delete

New Request: Create

Result: The new create, pending or immediate, replaces the pending delete.

Scenario-3

Pending Action: Create

New Request: Delete

Result: The new delete, pending or immediate, replaces the pending create.

Scenario-4

Pending Action: Create

New Request: Create

Result: The new create is added to the pending create.

Important: If the new create request is Apply Immediately - Yes, all creates are performed immediately. If the new create request is Apply Immediately - No, all creates are pending.

notification_topic_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the cache cluster owner.

notification_topic_status: Option<String>

The status of the Amazon SNS notification topic. Notifications are sent only if the status is active.

Valid values: active | inactive

num_cache_nodes: Option<i64>

The number of cache nodes that the cache cluster should have. If the value for NumCacheNodes is greater than the sum of the number of current cache nodes and the number of cache nodes pending creation (which may be zero), more nodes are added. If the value is less than the number of existing cache nodes, nodes are removed. If the value is equal to the number of current cache nodes, any pending add or remove requests are canceled.

If you are removing cache nodes, you must use the CacheNodeIdsToRemove parameter to provide the IDs of the specific cache nodes to remove.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

Adding or removing Memcached cache nodes can be applied immediately or as a pending operation (see ApplyImmediately).

A pending operation to modify the number of cache nodes in a cluster during its maintenance window, whether by adding or removing nodes in accordance with the scale out architecture, is not queued. The customer's latest request to add or remove nodes to the cluster overrides any previous pending operations to modify the number of cache nodes in the cluster. For example, a request to remove 2 nodes would override a previous pending operation to remove 3 nodes. Similarly, a request to add 2 nodes would override a previous pending operation to remove 3 nodes and vice versa. As Memcached cache nodes may now be provisioned in different Availability Zones with flexible cache node placement, a request to add nodes does not automatically override a previous pending operation to add nodes. The customer can modify the previous pending operation to add more nodes or explicitly cancel the pending request and retry the new request. To cancel pending operations to modify the number of cache nodes in a cluster, use the ModifyCacheCluster request and set NumCacheNodes equal to the number of cache nodes currently in the cache cluster.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

security_group_ids: Option<Vec<String>>

Specifies the VPC Security Groups associated with the cache cluster.

This parameter can be used only with clusters that are created in an Amazon Virtual Private Cloud (Amazon VPC).

snapshot_retention_limit: Option<i64>

The number of days for which ElastiCache retains automatic cache cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

snapshot_window: Option<String>

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your cache cluster.