Struct rusoto_elasticache::ModifyReplicationGroupShardConfigurationMessage

pub struct ModifyReplicationGroupShardConfigurationMessage {
    pub apply_immediately: bool,
    pub node_group_count: i64,
    pub node_groups_to_remove: Option<Vec<String>>,
    pub replication_group_id: String,
    pub resharding_configuration: Option<Vec<ReshardingConfiguration>>,
}

Represents the input for a ModifyReplicationGroupShardConfiguration operation.

Fields
apply_immediately: bool

Indicates that the shard reconfiguration process begins immediately. At present, the only permitted value for this parameter is true.

Value: true

node_group_count: i64

The number of node groups (shards) that results from the modification of the shard configuration.

node_groups_to_remove: Option<Vec<String>>

If the value of NodeGroupCount is less than the current number of node groups (shards), NodeGroupsToRemove is a required list of node group ids to remove from the cluster.

replication_group_id: String

The name of the Redis (cluster mode enabled) cluster (replication group) on which the shards are to be configured.

resharding_configuration: Option<Vec<ReshardingConfiguration>>

Specifies the preferred availability zones for each node group in the cluster. If the value of NodeGroupCount is greater than the current number of node groups (shards), you can use this parameter to specify the preferred availability zones of the cluster's shards. If you omit this parameter ElastiCache selects availability zones for you.

You can specify this parameter only if the value of NodeGroupCount is greater than the current number of node groups (shards).