Struct rusoto_elasticache::ReplicationGroup

pub struct ReplicationGroup {
    pub at_rest_encryption_enabled: Option<bool>,
    pub auth_token_enabled: Option<bool>,
    pub automatic_failover: Option<String>,
    pub cache_node_type: Option<String>,
    pub cluster_enabled: Option<bool>,
    pub configuration_endpoint: Option<Endpoint>,
    pub description: Option<String>,
    pub member_clusters: Option<Vec<String>>,
    pub node_groups: Option<Vec<NodeGroup>>,
    pub pending_modified_values: Option<ReplicationGroupPendingModifiedValues>,
    pub replication_group_id: Option<String>,
    pub snapshot_retention_limit: Option<i64>,
    pub snapshot_window: Option<String>,
    pub snapshotting_cluster_id: Option<String>,
    pub status: Option<String>,
    pub transit_encryption_enabled: Option<bool>,
}

Contains all of the attributes of a specific Redis replication group.

Fields
at_rest_encryption_enabled: Option<bool>

A flag that enables encryption at-rest when set to true.

You cannot modify the value of AtRestEncryptionEnabled after the cluster is created. To enable encryption at-rest on a cluster you must set AtRestEncryptionEnabled to true when you create a cluster.

Default: false

auth_token_enabled: Option<bool>

A flag that enables using an AuthToken (password) when issuing Redis commands.

Default: false

automatic_failover: Option<String>

Indicates the status of Multi-AZ with automatic failover for this Redis replication group.

Amazon ElastiCache for Redis does not support Multi-AZ with automatic failover on:

Redis versions earlier than 2.8.6.

Redis (cluster mode disabled): T1 and T2 cache node types.

Redis (cluster mode enabled): T1 node types.

cache_node_type: Option<String>

The name of the compute and memory capacity node type for each node in the replication group.

cluster_enabled: Option<bool>

A flag indicating whether or not this replication group is cluster enabled; i.e., whether its data can be partitioned across multiple shards (API/CLI: node groups).

Valid values: true | false

configuration_endpoint: Option<Endpoint>

The configuration endpoint for this replication group. Use the configuration endpoint to connect to this replication group.

description: Option<String>

The user supplied description of the replication group.

member_clusters: Option<Vec<String>>

The identifiers of all the nodes that are part of this replication group.

node_groups: Option<Vec<NodeGroup>>

A list of node groups in this replication group. For Redis (cluster mode disabled) replication groups, this is a single-element list. For Redis (cluster mode enabled) replication groups, the list contains an entry for each node group (shard).

pending_modified_values: Option<ReplicationGroupPendingModifiedValues>

A group of settings to be applied to the replication group, either immediately or during the next maintenance window.

replication_group_id: Option<String>

The identifier for the replication group.

snapshot_retention_limit: Option<i64>

The number of days for which ElastiCache retains automatic cluster snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

snapshot_window: Option<String>

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

snapshotting_cluster_id: Option<String>

The cluster ID that is used as the daily snapshot source for the replication group.

status: Option<String>

The current state of this replication group - creating, available, modifying, deleting, create-failed, snapshotting.

transit_encryption_enabled: Option<bool>

A flag that enables in-transit encryption when set to true.

You cannot modify the value of TransitEncryptionEnabled after the cluster is created. To enable in-transit encryption on a cluster you must set TransitEncryptionEnabled to true when you create a cluster.

Default: false