Struct rusoto_elasticache::Snapshot

pub struct Snapshot {
    pub auto_minor_version_upgrade: Option<bool>,
    pub automatic_failover: Option<String>,
    pub cache_cluster_create_time: Option<String>,
    pub cache_cluster_id: Option<String>,
    pub cache_node_type: Option<String>,
    pub cache_parameter_group_name: Option<String>,
    pub cache_subnet_group_name: Option<String>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub node_snapshots: Option<Vec<NodeSnapshot>>,
    pub num_cache_nodes: Option<i64>,
    pub num_node_groups: Option<i64>,
    pub port: Option<i64>,
    pub preferred_availability_zone: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub replication_group_description: Option<String>,
    pub replication_group_id: Option<String>,
    pub snapshot_name: Option<String>,
    pub snapshot_retention_limit: Option<i64>,
    pub snapshot_source: Option<String>,
    pub snapshot_status: Option<String>,
    pub snapshot_window: Option<String>,
    pub topic_arn: Option<String>,
    pub vpc_id: Option<String>,
}

Represents a copy of an entire Redis cache cluster as of the time when the snapshot was taken.

Fields
auto_minor_version_upgrade: Option<bool>

This parameter is currently disabled.

automatic_failover: Option<String>

Indicates the status of Multi-AZ for the source replication group.

ElastiCache Multi-AZ replication groups are not supported on:

Redis versions earlier than 2.8.6.

Redis (cluster mode disabled):T1 and T2 cache node types.

Redis (cluster mode enabled): T1 node types.

cache_cluster_create_time: Option<String>

The date and time when the source cache cluster was created.

cache_cluster_id: Option<String>

The user-supplied identifier of the source cache cluster.

cache_node_type: Option<String>

The name of the compute and memory capacity node type for the source cache cluster.

Valid node types are as follows:

General purpose:

Current generation: cache.t2.micro, cache.t2.small, cache.t2.medium, cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge, cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

Previous generation: cache.t1.micro, cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

Compute optimized: cache.c1.xlarge

Memory optimized:

Current generation: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Previous generation: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

Notes:

All T2 instances are created in an Amazon Virtual Private Cloud (Amazon VPC).

Redis backup/restore is not supported for Redis (cluster mode disabled) T1 and T2 instances. Backup/restore is supported on Redis (cluster mode enabled) T2 instances.

Redis Append-only files (AOF) functionality is not supported for T1 or T2 instances.

For a complete listing of node types and specifications, see Amazon ElastiCache Product Features and Details and either Cache Node Type-Specific Parameters for Memcached or Cache Node Type-Specific Parameters for Redis.

cache_parameter_group_name: Option<String>

The cache parameter group that is associated with the source cache cluster.

cache_subnet_group_name: Option<String>

The name of the cache subnet group associated with the source cache cluster.

engine: Option<String>

The name of the cache engine (memcached or redis) used by the source cache cluster.

engine_version: Option<String>

The version of the cache engine version that is used by the source cache cluster.

node_snapshots: Option<Vec<NodeSnapshot>>

A list of the cache nodes in the source cache cluster.

num_cache_nodes: Option<i64>

The number of cache nodes in the source cache cluster.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

num_node_groups: Option<i64>

The number of node groups (shards) in this snapshot. When restoring from a snapshot, the number of node groups (shards) in the snapshot and in the restored replication group must be the same.

port: Option<i64>

The port number used by each cache nodes in the source cache cluster.

preferred_availability_zone: Option<String>

The name of the Availability Zone in which the source cache cluster is located.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

replication_group_description: Option<String>

A description of the source replication group.

replication_group_id: Option<String>

The unique identifier of the source replication group.

snapshot_name: Option<String>

The name of a snapshot. For an automatic snapshot, the name is system-generated. For a manual snapshot, this is the user-provided name.

snapshot_retention_limit: Option<i64>

For an automatic snapshot, the number of days for which ElastiCache retains the snapshot before deleting it.

For manual snapshots, this field reflects the SnapshotRetentionLimit for the source cache cluster when the snapshot was created. This field is otherwise ignored: Manual snapshots do not expire, and can only be deleted using the DeleteSnapshot operation.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

snapshot_source: Option<String>

Indicates whether the snapshot is from an automatic backup (automated) or was created manually (manual).

snapshot_status: Option<String>

The status of the snapshot. Valid values: creating | available | restoring | copying | deleting.

snapshot_window: Option<String>

The daily time range during which ElastiCache takes daily snapshots of the source cache cluster.

topic_arn: Option<String>

The Amazon Resource Name (ARN) for the topic used by the source cache cluster for publishing notifications.

vpc_id: Option<String>

The Amazon Virtual Private Cloud identifier (VPC ID) of the cache subnet group for the source cache cluster.