Struct rusoto_elasticache::CreateCacheClusterMessage

pub struct CreateCacheClusterMessage {
    pub az_mode: Option<String>,
    pub auth_token: Option<String>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub cache_cluster_id: String,
    pub cache_node_type: Option<String>,
    pub cache_parameter_group_name: Option<String>,
    pub cache_security_group_names: Option<Vec<String>>,
    pub cache_subnet_group_name: Option<String>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub notification_topic_arn: Option<String>,
    pub num_cache_nodes: Option<i64>,
    pub port: Option<i64>,
    pub preferred_availability_zone: Option<String>,
    pub preferred_availability_zones: Option<Vec<String>>,
    pub preferred_maintenance_window: Option<String>,
    pub replication_group_id: Option<String>,
    pub security_group_ids: Option<Vec<String>>,
    pub snapshot_arns: Option<Vec<String>>,
    pub snapshot_name: Option<String>,
    pub snapshot_retention_limit: Option<i64>,
    pub snapshot_window: Option<String>,
    pub tags: Option<Vec<Tag>>,
}

Represents the input of a CreateCacheCluster operation.

Fields
az_mode: Option<String>

Specifies whether the nodes in this Memcached cluster are created in a single Availability Zone or created across multiple Availability Zones in the cluster's region.

This parameter is only supported for Memcached clusters.

If the AZMode and PreferredAvailabilityZones are not specified, ElastiCache assumes single-az mode.

auth_token: Option<String>

Reserved parameter. The password used to access a password protected server.

Password constraints:

Must be only printable ASCII characters.

Must be at least 16 characters and no more than 128 characters in length.

Cannot contain any of the following characters: '/', '"', or '@'.

For more information, see AUTH password at http://redis.io/commands/AUTH.

auto_minor_version_upgrade: Option<bool>

This parameter is currently disabled.

cache_cluster_id: String

The node group (shard) identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

cache_node_type: Option<String>

The compute and memory capacity of the nodes in the node group (shard).

The following node types are supported by ElastiCache. Generally speaking, the current generation types provide more memory and computational power at lower cost when compared to their equivalent previous generation counterparts.

General purpose:

Current generation:

M5 node types: cache.m5.large, cache.m5.xlarge, cache.m5.2xlarge, cache.m5.4xlarge, cache.m5.12xlarge, cache.m5.24xlarge

M4 node types: cache.m4.large, cache.m4.xlarge, cache.m4.2xlarge, cache.m4.4xlarge, cache.m4.10xlarge

T2 node types: cache.t2.micro, cache.t2.small, cache.t2.medium

Previous generation: (not recommended)

T1 node types: cache.t1.micro

M1 node types: cache.m1.small, cache.m1.medium, cache.m1.large, cache.m1.xlarge

M3 node types: cache.m3.medium, cache.m3.large, cache.m3.xlarge, cache.m3.2xlarge

Compute optimized:

Previous generation: (not recommended)

C1 node types: cache.c1.xlarge

Memory optimized:

Current generation:

R5 node types: cache.r5.large, cache.r5.xlarge, cache.r5.2xlarge, cache.r5.4xlarge, cache.r5.12xlarge, cache.r5.24xlarge

R4 node types: cache.r4.large, cache.r4.xlarge, cache.r4.2xlarge, cache.r4.4xlarge, cache.r4.8xlarge, cache.r4.16xlarge

Previous generation: (not recommended)

M2 node types: cache.m2.xlarge, cache.m2.2xlarge, cache.m2.4xlarge

R3 node types: cache.r3.large, cache.r3.xlarge, cache.r3.2xlarge, cache.r3.4xlarge, cache.r3.8xlarge

Additional node type info

All current generation instance types are created in Amazon VPC by default.

Redis append-only files (AOF) are not supported for T1 or T2 instances.

Redis Multi-AZ with automatic failover is not supported on T1 instances.

Redis configuration variables appendonly and appendfsync are not supported on Redis version 2.8.22 and later.

cache_parameter_group_name: Option<String>

The name of the parameter group to associate with this cluster. If this argument is omitted, the default parameter group for the specified engine is used. You cannot use any parameter group which has cluster-enabled='yes' when creating a cluster.

cache_security_group_names: Option<Vec<String>>

A list of security group names to associate with this cluster.

Use this parameter only when you are creating a cluster outside of an Amazon Virtual Private Cloud (Amazon VPC).

cache_subnet_group_name: Option<String>

The name of the subnet group to be used for the cluster.

Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (Amazon VPC).

If you're going to launch your cluster in an Amazon VPC, you need to create a subnet group before you start creating a cluster. For more information, see Subnets and Subnet Groups.

engine: Option<String>

The name of the cache engine to be used for this cluster.

Valid values for this parameter are: memcached | redis

engine_version: Option<String>

The version number of the cache engine to be used for this cluster. To view the supported cache engine versions, use the DescribeCacheEngineVersions operation.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing cluster or replication group and create it anew with the earlier engine version.

notification_topic_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic to which notifications are sent.

The Amazon SNS topic owner must be the same as the cluster owner.

num_cache_nodes: Option<i64>

The initial number of cache nodes that the cluster has.

For clusters running Redis, this value must be 1. For clusters running Memcached, this value must be between 1 and 20.

If you need more than 20 nodes for your Memcached cluster, please fill out the ElastiCache Limit Increase Request form at http://aws.amazon.com/contact-us/elasticache-node-limit-request/.

port: Option<i64>

The port number on which each of the cache nodes accepts connections.

preferred_availability_zone: Option<String>

The EC2 Availability Zone in which the cluster is created.

All nodes belonging to this Memcached cluster are placed in the preferred Availability Zone. If you want to create your nodes across multiple Availability Zones, use PreferredAvailabilityZones.

Default: System chosen Availability Zone.

preferred_availability_zones: Option<Vec<String>>

A list of the Availability Zones in which cache nodes are created. The order of the zones in the list is not important.

This option is only supported on Memcached.

If you are creating your cluster in an Amazon VPC (recommended) you can only locate nodes in Availability Zones that are associated with the subnets in the selected subnet group.

The number of Availability Zones listed must equal the value of NumCacheNodes.

If you want all the nodes in the same Availability Zone, use PreferredAvailabilityZone instead, or repeat the Availability Zone multiple times in the list.

Default: System chosen Availability Zones.

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

replication_group_id: Option<String>

The ID of the replication group to which this cluster should belong. If this parameter is specified, the cluster is added to the specified replication group as a read replica; otherwise, the cluster is a standalone primary that is not part of any replication group.

If the specified replication group is Multi-AZ enabled and the Availability Zone is not specified, the cluster is created in Availability Zones that provide the best spread of read replicas across Availability Zones.

This parameter is only valid if the Engine parameter is redis.

security_group_ids: Option<Vec<String>>

One or more VPC security groups associated with the cluster.

Use this parameter only when you are creating a cluster in an Amazon Virtual Private Cloud (Amazon VPC).

snapshot_arns: Option<Vec<String>>

A single-element string list containing an Amazon Resource Name (ARN) that uniquely identifies a Redis RDB snapshot file stored in Amazon S3. The snapshot file is used to populate the node group (shard). The Amazon S3 object name in the ARN cannot contain any commas.

This parameter is only valid if the Engine parameter is redis.

Example of an Amazon S3 ARN: arn:aws:s3:::my_bucket/snapshot1.rdb

snapshot_name: Option<String>

The name of a Redis snapshot from which to restore data into the new node group (shard). The snapshot status changes to restoring while the new node group (shard) is being created.

This parameter is only valid if the Engine parameter is redis.

snapshot_retention_limit: Option<i64>

The number of days for which ElastiCache retains automatic snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot taken today is retained for 5 days before being deleted.

This parameter is only valid if the Engine parameter is redis.

Default: 0 (i.e., automatic backups are disabled for this cache cluster).

snapshot_window: Option<String>

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of your node group (shard).

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

This parameter is only valid if the Engine parameter is redis.

tags: Option<Vec<Tag>>

A list of cost allocation tags to be added to this resource.