Struct rusoto_elasticache::ModifyReplicationGroupMessage

pub struct ModifyReplicationGroupMessage {
    pub apply_immediately: Option<bool>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub automatic_failover_enabled: Option<bool>,
    pub cache_node_type: Option<String>,
    pub cache_parameter_group_name: Option<String>,
    pub cache_security_group_names: Option<Vec<String>>,
    pub engine_version: Option<String>,
    pub notification_topic_arn: Option<String>,
    pub notification_topic_status: Option<String>,
    pub preferred_maintenance_window: Option<String>,
    pub primary_cluster_id: Option<String>,
    pub replication_group_description: Option<String>,
    pub replication_group_id: String,
    pub security_group_ids: Option<Vec<String>>,
    pub snapshot_retention_limit: Option<i64>,
    pub snapshot_window: Option<String>,
    pub snapshotting_cluster_id: Option<String>,
}

Represents the input of a ModifyReplicationGroups operation.

Fields
apply_immediately: Option<bool>

If true, this parameter causes the modifications in this request and any pending modifications to be applied, asynchronously and as soon as possible, regardless of the PreferredMaintenanceWindow setting for the replication group.

If false, changes to the nodes in the replication group are applied on the next maintenance reboot, or the next failure reboot, whichever occurs first.

Valid values: true | false

Default: false

auto_minor_version_upgrade: Option<bool>

This parameter is currently disabled.

automatic_failover_enabled: Option<bool>

Determines whether a read replica is automatically promoted to read/write primary if the existing primary encounters a failure.

Valid values: true | false

Amazon ElastiCache for Redis does not support Multi-AZ with automatic failover on:

Redis versions earlier than 2.8.6.

Redis (cluster mode disabled): T1 node types.

Redis (cluster mode enabled): T1 node types.

cache_node_type: Option<String>

A valid cache node type that you want to scale this replication group to.

cache_parameter_group_name: Option<String>

The name of the cache parameter group to apply to all of the clusters in this replication group. This change is asynchronously applied as soon as possible for parameters when the ApplyImmediately parameter is specified as true for this request.

cache_security_group_names: Option<Vec<String>>

A list of cache security group names to authorize for the clusters in this replication group. This change is asynchronously applied as soon as possible.

This parameter can be used only with replication group containing clusters running outside of an Amazon Virtual Private Cloud (Amazon VPC).

Constraints: Must contain no more than 255 alphanumeric characters. Must not be Default.

engine_version: Option<String>

The upgraded version of the cache engine to be run on the clusters in the replication group.

Important: You can upgrade to a newer engine version (see Selecting a Cache Engine and Version), but you cannot downgrade to an earlier engine version. If you want to use an earlier engine version, you must delete the existing replication group and create it anew with the earlier engine version.

notification_topic_arn: Option<String>

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications are sent.

The Amazon SNS topic owner must be same as the replication group owner.

notification_topic_status: Option<String>

The status of the Amazon SNS notification topic for the replication group. Notifications are sent only if the status is active.

Valid values: active | inactive

preferred_maintenance_window: Option<String>

Specifies the weekly time range during which maintenance on the cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period.

Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:23:00-mon:01:30

primary_cluster_id: Option<String>

For replication groups with a single primary, if this parameter is specified, ElastiCache promotes the specified cluster in the specified replication group to the primary role. The nodes of all other clusters in the replication group are read replicas.

replication_group_description: Option<String>

A description for the replication group. Maximum length is 255 characters.

replication_group_id: String

The identifier of the replication group to modify.

security_group_ids: Option<Vec<String>>

Specifies the VPC Security Groups associated with the clusters in the replication group.

This parameter can be used only with replication group containing clusters running in an Amazon Virtual Private Cloud (Amazon VPC).

snapshot_retention_limit: Option<i64>

The number of days for which ElastiCache retains automatic node group (shard) snapshots before deleting them. For example, if you set SnapshotRetentionLimit to 5, a snapshot that was taken today is retained for 5 days before being deleted.

Important If the value of SnapshotRetentionLimit is set to zero (0), backups are turned off.

snapshot_window: Option<String>

The daily time range (in UTC) during which ElastiCache begins taking a daily snapshot of the node group (shard) specified by SnapshottingClusterId.

Example: 05:00-09:00

If you do not specify this parameter, ElastiCache automatically chooses an appropriate time range.

snapshotting_cluster_id: Option<String>

The cluster ID that is used as the daily snapshot source for the replication group. This parameter cannot be set for Redis (cluster mode enabled) replication groups.